fix: mako theme replacement (#8)
* Update Theme on Mako config

Instead of replacing the mako config file each time this updates just the colors.  It only copies the template file if no config file exist.

* Update 92-enable-mako-theme.conf

* sed on two line too slow

when the system does the two sed commands sometimes it doesn't get the the second before mako tries to read the config.  By putting them on the same line both with run as the same time.