feat: dynamic theming based on sway vars and GTK (#26)
* Rename colors to match pywal template

* Delete matcha-blue theme, maintenance burden

* Edit sway modes to use new color template

* Delete kitty template, kitty isn't used anymore

* Remove unnecessary gtk settings.ini changes

* Disable Qt theme changes for now (manjaro has their default)

* Remove unnecessary gtk templates (can be done via gsettings)

* Remove unnecessary mako template, can be done via sway vars

* Remove unnecessary swaylock config, can be done via sway vars

* Remove unnecessary .css theme (waybar|wlogout), can be done via gtk theme vars

* Refactor matcha-green theme definitions

* Use GTK theme colors in Waybar

* Use GTK theme colors in wlogout

* Remove script for mako theming, now done via CLI arguments

* Delete script for linking .css themes, GTK theme is used instead

* Fix wob accent color

* Add variable for notification daemon. Now using CLI args for theming

* Start $notifications instead of static mako command

* Fix mako dynamic theming

Apparently the previous command didn't work for several reasons
1. --margin 20% was incorrect, it should be --margin 20
2. Font and colors were used incorrectly (only first argument was read, sway makes no assumptions and thus doesn't sanitize vars by default), and mako is intolerant to such arguments

* Make rofi font dynamic

Was set statically before. With Default $gui-font value ('Roboto 11') it looks a bit small, if there are no reasons against, it would probably make sense to change it to 'Roboto 14' for consistency.

* fix help on waybar

The " was in the wrong spot and it broke the help.

* Rofi theme

To use the theme vars you have to include them before.  
You were missing the # in front of each of the colors in your method of changing colors and I was getting the wrong colors.
I think I changed it to the new theming method as per: https://github.com/davatorium/rofi/blob/next/doc/rofi.1.markdown#layout-and-theming

* Mako theming

Now that the theme is included mako almost worked.  It still needed # in front of the colors.
Also the notify stayed forever.  We should make it time out.

* Fix additional variables in stock theme definitions file

* undo changes to rofi and mako for color

undo some changes after change to theme file.

* add include

include your theme variables

* move theme var

Moved theme include out of definitions

* moved theme include

Moved theme include to main config.

* Java fix

fix java programs.

* done in another commit

done in another pull

* Add support for wob >= 0.22

* Delete trailing spaces


Co-authored-by: FallenChromium <fallenchromium@gmail.com>
Co-authored-by: Mathew <44036272+Mathew-D@users.noreply.github.com>
Co-authored-by: Jonas Strassel <info@jonas-strassel.de>